+++ util/buildComponents.js
@@ -8,6 +8,9 @@ import {
   has,
   forEach,
   mergeWith,
+  keys,
+  values,
+  isString,
 } from "lodash-es";
 
 // Reasons for reservation:
@@ -23,6 +26,10 @@ import {
 // - All the other reserved keys added in 0.7.0 are used by the game's Artitas
 //   serializer. Since objects are no longer the only acceptable argument
 //   to a builder, I have to guard against all of them now (or try to).
+// - $none specifies that a field explicitly defaults to "undefined".
+//   This is distinct from not specifying a default, in that
+//   it allows a multi-arg builder to be conditionally treated as a
+//   single-arg builder.
 const RESERVED = [
   "$content",
   "$args",
@@ -35,6 +42,7 @@ const RESERVED = [
   "$valuereference",
   "$type",
   "$t",
+  "$none",
 ];
 const PREFIX = "$";
 
@@ -65,9 +73,15 @@ export function importBuilders(source, registry) {
       result = false;
     } else
       forEach(definition.$args, (arg) => {
+        // Arg has a default value, sanity-check its name
+        // instead of the whole entry
+        if (isPlainObject(arg) && keys(arg).length === 1) {
+          arg = keys(arg)[0];
+        }
+
         if (typeof arg !== "string" || !arg.startsWith(PREFIX)) {
           console.error(
-            `Builder "${name}": $args must contain only strings starting with "${PREFIX}"!`,
+            `Builder "${name}": $args must contain only strings or single-key objects whose key starts with "${PREFIX}"!`,
           );
           result = false;
         } else if (RESERVED.includes(arg)) {
@@ -95,6 +109,24 @@ function safeMerge(objValue, srcValue) {
   return undefined;
 }
 
+// Resolves an argument to either a user-provided value or a default
+// (which may itself be a different user-provided value)
+function getValueOrDefault(userParams, arg, $args, argKeys, index) {
+  return (
+    userParams[arg] ?? getDefaultValue(userParams, arg, $args, argKeys, index)
+  );
+}
+
+function getDefaultValue(userParams, arg, $args, argKeys, index) {
+  if (index === undefined) index = argKeys.findIndex((key) => key === arg);
+  let defaultValue =
+    typeof $args[index] === "object" ? values($args[index])[0] : undefined;
+  if (defaultValue === "$none") defaultValue = undefined;
+  else if (argKeys.includes(defaultValue))
+    defaultValue = getValueOrDefault(userParams, defaultValue, $args, argKeys);
+  return defaultValue;
+}
+
 export function evaluateBuilders(data, builders) {
   if (!builders) return data;
 
@@ -115,11 +147,24 @@ export function evaluateBuilders(data, builders) {
         let userParams = value;
         const { $args, ...template } = builder;
 
+        // This is a bit of a misnomer - XenoYAML does not,
+        // and does not *want to*, check whether all the args are defined...
+        // but in order to enable shorthand invocation on a multi-arg builder,
+        // no more than one arg can be defined without a default value.
+        // For args that can literally be undefined, this default value should be
+        // the reserved keyword `$none`.
+        const requiredArgs =
+          $args?.length > 1 ? ($args || []).filter(isString) : $args;
+        const argKeys =
+          requiredArgs === $args
+            ? $args
+            : $args.map((arg) => (isString(arg) ? arg : keys(arg)[0]));
+
         // Shorthand invocation:
         // If the argument map isn't a map (or doesn't contain the argument key),
         // then it must be the argument itself!
-        if ($args?.length === 1) {
-          const singleArgName = $args[0];
+        if (requiredArgs?.length === 1) {
+          const singleArgName = requiredArgs[0];
 
           if (userParams?.$noArg) {
             delete userParams.$noArg;
@@ -129,17 +174,20 @@ export function evaluateBuilders(data, builders) {
           ) {
             userParams = { [singleArgName]: userParams };
           }
-        } else if ($args?.length && !isPlainObject(userParams)) {
+        } else if (requiredArgs?.length && !isPlainObject(userParams)) {
           throw new SyntaxError(
-            `Attempted to pass a scalar or array to a multi-arg builder! ${key} received the payload "${userParams}"`,
+            `Attempted to pass a scalar or array to a builder expecting multiple arguments! ${key} received the payload "${userParams}"`,
           );
         }
 
         // Map arguments (unbound become undefined)
         const argValues = {};
-        if ($args) {
-          $args.forEach((arg) => {
-            argValues[arg] = evaluateBuilders(userParams[arg], builders);
+        if (argKeys) {
+          argKeys.forEach((arg, index) => {
+            argValues[arg] = evaluateBuilders(
+              getValueOrDefault(userParams, arg, $args, argKeys, index),
+              builders,
+            );
           });
         }
 
@@ -151,7 +199,7 @@ export function evaluateBuilders(data, builders) {
         );
 
         // Process overrides (the non-arg siblings in the call)
-        const rawOverrides = omit(userParams, $args || []);
+        const rawOverrides = omit(userParams, argKeys || []);
         const cleanOverrides = evaluateBuilders(rawOverrides, builders);
 
         // Merge the builder result into our result collector